Move event struct into separate file.
